Задача 1 к Уроку 2 исправление

+++ src/main/kotlin/KotlinSprint_Lesson2/Task_1.kt
@@ -1,12 +1,13 @@
 package org.example.KotlinSprint_Lesson2
 
 fun main() {
-    val numberOfStudents = 4.00f
-    val gradeStudent1 = 3.00f
-    val gradeStudent2 = 4.00f
-    val gradeStudent3 = 3.00f
-    val gradeStudent4 = 5.00f
-    val averageScore = ((gradeStudent1 + gradeStudent2 + gradeStudent3 + gradeStudent4) / numberOfStudents)
+    val numberOfStudents = 4
+    val gradeStudent1 = 3
+    val gradeStudent2 = 4
+    val gradeStudent3 = 3
+    val gradeStudent4 = 5
+    val averageScore =
+        ((gradeStudent1.toFloat() + gradeStudent2.toFloat() + gradeStudent3.toFloat() + gradeStudent4.toFloat()) / numberOfStudents)
 
     println("Средний бал по английскому языку - $averageScore")
 }
